Add int branch to binsearch
The binsearch C function now branches on double and integer types, so
no conversion is done if 'key' and 'vec' are the same type. Cast both
to double if they're not the same type.

Remove the bound() function and call lower_bound() and upper_bound()
directly from binsearch(). Also move the edge case checks into the
respective *_bound() functions, since we need to do the checks inside
the switch() statement.

Removing the as.double() call means ISO-8601 range-based subsetting
is nearly twice as fast for integer vectors and keys.

See #100.