Add int branch to binsearch
The binsearch C function now branches on double and integer types, so
no conversion is done if 'key' and 'vec' are the same type. Cast both
to double if they're not the same type.

Create a struct that contains int and double fields, so we can pass
it to a generic compare function. Create 4 compare functions, one for
each type/start combination. All these components are static inline,
with the hope the compiler can optimize them.

Removing the as.double() call means ISO-8601 range-based subsetting
is nearly twice as fast for integer vectors and keys.

See #100.